refacotr: positiveNumber 이름 수정 및 계산 PostiveNumber 에서 하도록 수정
