--- src/main/java/calculator/domain/Operand.java
@@ -1,15 +1,0 @@
-package calculator.domain;
-
-public class Operand {
-
-    private final int number;
-
-    public Operand(final int number) {
-        this.number = number;
-    }
-
-    public int getNumber() {
-        return number;
-    }
-
-}
